split the user interface
split to userBase and userFull for better enforcement control
