+++ apps/backend/src/app/users/repository/user.schema.ts
@@ -1,12 +1,12 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { Document } from 'mongoose';
-import { UserBase } from 'common/src/lib/Interfaces/user.interface';
 import { Role } from 'common/src/lib/Enums/role.enum';
+import { UserFull } from 'common/src/lib/Interfaces/user-full.interface';
 
 export type UserDocument = UserSchema & Document;
 
 @Schema({ collection: 'users' })
-export class UserSchema implements UserBase {
+export class UserSchema implements UserFull {
   @Prop({ required: true })
   username: string;
