+++ apps/backend/src/app/users/repository/users-repository.interface.ts
@@ -1,8 +1,10 @@
-import { UserBase } from 'common/src/lib/Interfaces/user.interface';
+import { UserBase } from 'common/src/lib/Interfaces/user-base.interface';
+import { UserFull } from 'common/src/lib/Interfaces/user-full.interface';
 
 export const USERS_REPOSITORY = Symbol('USERS_REPOSITORY');
 
 export interface IUsersRepository {
-  createUser(user: Partial<UserBase>): Promise<UserBase>;
-  getAllUsers(): Promise<{username: string}>;
+  findUserByEmail(email: string): Promise<UserFull | null>;
+  createUser(user: Partial<UserFull>): Promise<UserBase>;
+  getAllUsers(): Promise<UserBase[]>;
 }
