+++ common/src/lib/Interfaces/user-base.interface.ts
@@ -1,0 +1,5 @@
+
+export interface UserBase {
+  username: string;
+  email: string;
+}
