+++ common/src/lib/Interfaces/user-full.interface.ts
@@ -1,0 +1,7 @@
+import { Role } from '../Enums/role.enum';
+import { UserBase } from './user-base.interface';
+
+export interface UserFull extends UserBase {
+    password: string;
+    role: Role;
+}
