--- common/src/lib/Interfaces/user.interface.ts
@@ -1,8 +1,0 @@
-import { Role } from '../Enums/role.enum';
-
-export interface UserBase {
-  username: string;
-  email: string;
-  password: string;
-  role: Role;
-}
